+++ packages/connectanum_router/test/hook/build_hook_test.dart
@@ -141,7 +141,7 @@ void main() {
       addTearDown(() => packageRoot.delete(recursive: true));
       File('${packageRoot.path}/pubspec.yaml').writeAsStringSync('''
 name: connectanum_router
-version: 3.0.0-beta.3
+version: 3.0.0-beta.4
 environment:
   sdk: ^3.9.2
 ''');
@@ -186,7 +186,7 @@ environment:
               downloaded.map((uri) => uri.toString()),
               contains(
                 'https://github.com/konsultaner/connectanum-dart/releases/download/'
-                'v3.0.0-beta.3/${_releaseArchiveName()}',
+                'v3.0.0-beta.4/${_releaseArchiveName()}',
               ),
             );
           },
@@ -212,7 +212,7 @@ environment:
     )..createSync(recursive: true);
     File('${packageRoot.path}/pubspec.yaml').writeAsStringSync('''
 name: connectanum_router
-version: 3.0.0-beta.3
+version: 3.0.0-beta.4
 environment:
   sdk: ^3.9.2
 ''');